+++ src/FileBufferIO.c
@@ -548,6 +548,7 @@ ret0:			ret = 0;
 	fp->lastl = fp->lpReadPointer;
 	fp->lpReadPointer = fp->firstl;
 	doctypes_reassignDocumentTypeDescriptor(fp);
+	fp->vcStatus = gi_getStatus(fp->fname);
 
 readerr:
 	if (fd > 0) {